Rename: client/src/components/common/WriteButtonVer2.tsx -> client/src/components/common/NewPostButton.tsx
