+++ client/src/pages/event/Event.tsx
@@ -4,7 +4,7 @@ import { useIntersectionObserver } from "../../hooks/useIntersectionObserver";
 import PostEmpty from "../../components/communityAndEvent/PostEmpty";
 import PostList from "../../components/communityAndEvent/PostList";
 import LoadingCat from "../../components/loading/LoadingCat";
-import NewPostButton from "../../components/common/WriteButtonVer2";
+import NewPostButton from "../../components/common/NewPostButton";
 
 const Event = () => {
   const { data, isLoading, fetchNextPage, hasNextPage, isFetchingNextPage, isEmpty } = useEvents("views");
